Add category & some updated in templates

+++ books/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render,redirect
 from .forms import BookForm
 from django.http import HttpResponse
 from .models import Book
+from django.db.models import Count
 # Create your views here.
 
 def index(request):
@@ -22,17 +23,27 @@ def create(request):
 
 def edit(request,id):
 
-    book = Book.objects.get(pk=id)
-    form = BookForm(request.POST or None , instance=book)
-    if form.is_valid():
-        form.save()
-        return redirect("index")
-    return render(request,"books/edit.html",{
-        "form" : form,
-        "book" : book
-    })
+    
+    exists = Book.objects.filter(id=id).exists()
+    if exists:
+    
+        book = Book.objects.get(pk=id)
+        form = BookForm(request.POST or None , instance=book)
+        if form.is_valid():
+            form.save()
+            return redirect("index")
+        return render(request,"books/edit.html",{
+            "form" : form,
+            "book" : book
+        })
+
+    return redirect("index")
 
 def delete(request,id):
-    book = Book.objects.get(pk=id)
-    book.delete()
+    exists = Book.objects.filter(id=id).exists()
+    if exists:
+    
+        book = Book.objects.get(pk=id)
+        book.delete()
+        return redirect("index")
     return redirect("index")
